+++ src/interfaces/IERC721X.sol
@@ -3,6 +3,6 @@
 pragma solidity >=0.8.7 <0.9.0;
 
 interface IERC721X {
-    function originChainId() external view returns (uint16);
+    function originChainId() external view returns (uint32);
     function originAddress() external view returns (address);
 }
